requirements.txt
# Core data processing
pandas==2.0.3
numpy==1.24.3
requests==2.31.0

# Database connections
psycopg2-binary==2.9.7
sqlalchemy==2.0.19

# Airflow (using same version as base image)
# apache-airflow[postgres]==2.7.1  # Don't install again, use base image

# dbt (will be installed separately in Dockerfile)
# dbt-postgres==1.6.0

# Dashboard
streamlit==1.28.0
plotly==5.15.0

# Data science
scikit-learn==1.3.0
textblob==0.17.1

# Development tools
jupyter==1.0.0
black==23.7.0
pylint==2.17.5

# GCP (for later deployment)
google-cloud-bigquery==3.11.4
google-cloud-storage==2.10.0

# Additional dependencies for data generation
faker==19.6.2
python-dateutil==2.8.2